cleanup base-docker and deployments
