+++ generators/base-docker/docker-base.mjs
@@ -26,7 +26,7 @@ import { GENERATOR_JHIPSTER } from '../generator-constants.mjs';
 const { MAVEN } = buildToolTypes;
 const { MONOLITH, MICROSERVICE, GATEWAY } = applicationTypes;
 
-export { checkDocker } from './docker-utils.mjs';
+export { checkDocker } from '../docker/support/index.mjs';
 
 /**
  * Check Images
@@ -71,18 +71,6 @@ export function configureImageNames() {
     const originalImageName = this.appConfigs[i].baseName.toLowerCase();
     const targetImageName = this.dockerRepositoryName ? `${this.dockerRepositoryName}/${originalImageName}` : originalImageName;
     this.appConfigs[i].targetImageName = targetImageName;
-  }
-}
-
-/**
- * Set Apps Folder Paths
- */
-export function setAppsFolderPaths() {
-  if (this.applicationType) return;
-  this.appsFolderPaths = [];
-  for (let i = 0; i < this.appsFolders.length; i++) {
-    const path = this.destinationPath(this.directoryPath + this.appsFolders[i]);
-    this.appsFolderPaths.push(path);
   }
 }
