+++ generators/base-docker/generator.mjs
@@ -19,14 +19,12 @@
 import _ from 'lodash';
 
 import prompts from './docker-prompts.mjs';
-import BlueprintBaseGenerator from '../base/index.mjs';
+import BaseWorkspacesGenerator from '../base-workspaces/index.mjs';
 import { GENERATOR_DOCKER_COMPOSE } from '../generator-list.mjs';
-import { loadFromYoRc, checkDocker, checkImages, generateJwtSecret, setAppsFolderPaths } from './docker-base.mjs';
+import { loadFromYoRc, checkDocker, checkImages, generateJwtSecret } from './docker-base.mjs';
 import statistics from '../statistics.mjs';
 import { applicationOptions, deploymentOptions } from '../../jdl/jhipster/index.mjs';
-
-import { dockerContainers as elasticDockerContainer } from '../generator-constants.mjs';
-import { dockerPlaceholderGenerator, getDockerfileContainers } from '../docker/utils.mjs';
+import { loadDockerDependenciesTask } from '../base-workspaces/internal/index.mjs';
 
 const { OptionNames } = applicationOptions;
 const { Options: DeploymentOptions } = deploymentOptions;
@@ -35,9 +33,9 @@ const { JWT_SECRET_KEY } = OptionNames;
 
 /**
  * @class
- * @extends {BlueprintBaseGenerator}
+ * @extends {BaseWorkspacesGenerator}
  */
-export default class BaseDockerGenerator extends BlueprintBaseGenerator {
+export default class BaseDockerGenerator extends BaseWorkspacesGenerator {
   constructor(args, options, features) {
     super(args, options, features);
 
@@ -52,22 +50,9 @@ export default class BaseDockerGenerator extends BlueprintBaseGenerator {
 
   get initializing() {
     return {
-      async setupServerConsts() {
-        const dockerfile = this.readTemplate(this.jhipsterTemplatePath('../../server/resources/Dockerfile'));
-        this.dockerContainers = this.prepareDependencies(
-          {
-            ...elasticDockerContainer,
-            ...getDockerfileContainers(dockerfile),
-          },
-          dockerPlaceholderGenerator,
-        );
-      },
-
+      loadDockerDependenciesTask,
       checkDocker,
-
-      loadConfig() {
-        loadFromYoRc.call(this);
-      },
+      loadFromYoRc,
     };
   }
 
@@ -93,7 +78,6 @@ export default class BaseDockerGenerator extends BlueprintBaseGenerator {
 
       checkImages,
       generateJwtSecret,
-      setAppsFolderPaths,
     };
   }
