+++ generators/kubernetes-helm/generator.mjs
@@ -20,12 +20,12 @@
 import fs from 'fs';
 import chalk from 'chalk';
 
-import BaseDockerGenerator from '../base-docker/index.mjs';
+import BaseWorkspacesGenerator from '../base-docker/index.mjs';
 
 import prompts from '../kubernetes/prompts.mjs';
 import { writeFiles } from './files.mjs';
 import { GENERATOR_KUBERNETES_HELM } from '../generator-list.mjs';
-import { checkImages, generateJwtSecret, configureImageNames, setAppsFolderPaths } from '../base-docker/docker-base.mjs';
+import { checkImages, generateJwtSecret, configureImageNames } from '../base-docker/docker-base.mjs';
 import {
   checkKubernetes,
   checkHelm,
@@ -38,14 +38,16 @@ import {
 import statistics from '../statistics.mjs';
 import { messageBrokerTypes } from '../../jdl/jhipster/index.mjs';
 import { getJdbcUrl, getR2dbcUrl } from '../spring-data-relational/support/index.mjs';
+import { loadDockerDependenciesTask } from '../base-workspaces/internal/index.mjs';
+import { checkDocker } from '../docker/support/index.mjs';
 
 const { KAFKA } = messageBrokerTypes;
 
 /**
  * @class
- * @extends {BaseDockerGenerator}
+ * @extends {BaseWorkspacesGenerator}
  */
-export default class KubernetesHelmGenerator extends BaseDockerGenerator {
+export default class KubernetesHelmGenerator extends BaseWorkspacesGenerator {
   async beforeQueue() {
     if (!this.fromBlueprint) {
       await this.composeWithBlueprints(GENERATOR_KUBERNETES_HELM);
@@ -58,7 +60,8 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
         this.log.log(chalk.white(`${chalk.bold('⎈')} Welcome to the JHipster Kubernetes Helm Generator ${chalk.bold('⎈')}`));
         this.log.log(chalk.white(`Files will be generated in folder: ${chalk.yellow(this.destinationRoot())}`));
       },
-      ...super.initializing,
+      loadDockerDependenciesTask,
+      checkDocker,
       checkKubernetes,
       checkHelm,
       loadConfig,
@@ -67,7 +70,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.INITIALIZING]() {
+  get [BaseWorkspacesGenerator.INITIALIZING]() {
     return this.delegateTasksToBlueprint(() => this.initializing);
   }
 
@@ -90,7 +93,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.PROMPTING]() {
+  get [BaseWorkspacesGenerator.PROMPTING]() {
     return this.delegateTasksToBlueprint(() => this.prompting);
   }
 
@@ -103,7 +106,6 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
       checkImages,
       generateJwtSecret,
       configureImageNames,
-      setAppsFolderPaths,
 
       setPostPromptProp() {
         this.appConfigs.forEach(element => {
@@ -118,7 +120,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.CONFIGURING]() {
+  get [BaseWorkspacesGenerator.CONFIGURING]() {
     return this.delegateTasksToBlueprint(() => this.configuring);
   }
 
@@ -138,7 +140,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.LOADING]() {
+  get [BaseWorkspacesGenerator.LOADING]() {
     return this.delegateTasksToBlueprint(() => this.loading);
   }
 
@@ -146,7 +148,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     return writeFiles();
   }
 
-  get [BaseDockerGenerator.WRITING]() {
+  get [BaseWorkspacesGenerator.WRITING]() {
     return this.delegateTasksToBlueprint(() => this.writing);
   }
 
@@ -188,7 +190,7 @@ export default class KubernetesHelmGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.END]() {
+  get [BaseWorkspacesGenerator.END]() {
     return this.delegateTasksToBlueprint(() => this.end);
   }
