+++ generators/kubernetes-knative/generator.mjs
@@ -22,12 +22,12 @@ import chalk from 'chalk';
 import shelljs from 'shelljs';
 import runAsync from 'run-async';
 
-import BaseDockerGenerator from '../base-docker/index.mjs';
+import BaseWorkspacesGenerator from '../base-docker/index.mjs';
 
 import prompts from './prompts.mjs';
 import { writeFiles } from './files.mjs';
 import { GENERATOR_KUBERNETES_KNATIVE } from '../generator-list.mjs';
-import { checkImages, generateJwtSecret, configureImageNames, setAppsFolderPaths } from '../base-docker/docker-base.mjs';
+import { checkImages, generateJwtSecret, configureImageNames } from '../base-docker/docker-base.mjs';
 import {
   checkHelm,
   checkKubernetes,
@@ -40,6 +40,8 @@ import {
 import statistics from '../statistics.mjs';
 import { kubernetesPlatformTypes, buildToolTypes, messageBrokerTypes } from '../../jdl/jhipster/index.mjs';
 import { getJdbcUrl } from '../spring-data-relational/support/index.mjs';
+import { loadDockerDependenciesTask } from '../base-workspaces/internal/index.mjs';
+import { checkDocker } from '../docker/support/index.mjs';
 
 const { GeneratorTypes } = kubernetesPlatformTypes;
 const { MAVEN } = buildToolTypes;
@@ -49,9 +51,9 @@ const { K8S } = GeneratorTypes;
 
 /**
  * @class
- * @extends {BaseDockerGenerator}
+ * @extends {BaseWorkspacesGenerator}
  */
-export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
+export default class KubernetesKnativeGenerator extends BaseWorkspacesGenerator {
   async beforeQueue() {
     if (!this.fromBlueprint) {
       await this.composeWithBlueprints(GENERATOR_KUBERNETES_KNATIVE);
@@ -64,7 +66,8 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
         this.log.log(chalk.white(`${chalk.bold('☸')} Welcome to the JHipster Kubernetes Knative Generator ${chalk.bold('☸')}`));
         this.log.log(chalk.white(`Files will be generated in the folder: ${chalk.yellow(this.destinationRoot())}`));
       },
-      ...super.initializing,
+      loadDockerDependenciesTask,
+      checkDocker,
       checkKubernetes,
       checkHelm,
       checkKnative: runAsync(function () {
@@ -94,7 +97,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.INITIALIZING]() {
+  get [BaseWorkspacesGenerator.INITIALIZING]() {
     return this.delegateTasksToBlueprint(() => this.initializing);
   }
 
@@ -114,7 +117,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.PROMPTING]() {
+  get [BaseWorkspacesGenerator.PROMPTING]() {
     return this.delegateTasksToBlueprint(() => this.prompting);
   }
 
@@ -127,7 +130,6 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
       checkImages,
       generateJwtSecret,
       configureImageNames,
-      setAppsFolderPaths,
 
       setPostPromptProp() {
         this.appConfigs.forEach(element => {
@@ -142,7 +144,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.CONFIGURING]() {
+  get [BaseWorkspacesGenerator.CONFIGURING]() {
     return this.delegateTasksToBlueprint(() => this.configuring);
   }
 
@@ -162,7 +164,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.LOADING]() {
+  get [BaseWorkspacesGenerator.LOADING]() {
     return this.delegateTasksToBlueprint(() => this.loading);
   }
 
@@ -170,7 +172,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     return writeFiles();
   }
 
-  get [BaseDockerGenerator.WRITING]() {
+  get [BaseWorkspacesGenerator.WRITING]() {
     return this.delegateTasksToBlueprint(() => this.writing);
   }
 
@@ -239,7 +241,7 @@ export default class KubernetesKnativeGenerator extends BaseDockerGenerator {
     };
   }
 
-  get [BaseDockerGenerator.END]() {
+  get [BaseWorkspacesGenerator.END]() {
     return this.delegateTasksToBlueprint(() => this.end);
   }
